chore: move multivalue to internals
