+++ src/components/internals/multivalue/index.js
@@ -1,0 +1,1 @@
+export { default } from './multivalue';
